docs: cleanup child.component.ts field documentation
