+++ src/test/angular-app/src/app/parent.directive.ts
@@ -2,5 +2,9 @@ import { Directive, Input } from "@angular/core";
 
 @Directive()
 export abstract class ParentDirective {
+    /**
+     * This is an input in the parent directive. It should also be
+     * visible in the docs of child classes
+     */
     @Input() parentInput?: string;
 }
